Usavršena metoda za dohvvat knjiga prema ratingu.

+++ FindAndRead/FindAndRead/Models/BooksByRatingData.cs
@@ -8,7 +8,7 @@ namespace FindAndRead.Models
     public class BooksByRatingData
     {
         public int brojCitanja { get; set; }
-        public float prosjecnaOcjena { get; set; }
+        public double prosjecnaOcjena { get; set; }
         public IEnumerable<ProcitanoVeza> listaCitanja { get; set; }
         public Book knjiga { get; set; }
     }
